% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_customRatioBuilder.R
\name{customRatioBuilder}
\alias{customRatioBuilder}
\title{Compute and Plot Custom Ratio of User's Choice Over Selected Years}
\usage{
customRatioBuilder(dataARL, numerator, denominator, members, years = NA)
}
\arguments{
\item{dataARL}{A dataframe containing ARL survey data directly
downloaded from ARL platform. The years should be placed along
rows. The first column must be 'Year', followed by other variables
in no particular order, e.g., 'Institution Name', 'Institution type',
etc. To download data from ARL Data Portal, it is recommended that
all variables are selected, with columns being 'Variables' and data
sorted by 'Institution Name' (default options).}

\item{numerator}{A character vector specifying a numeric statistic from the
ARL survey data to be used as the numerator. The vector should be
an exact match to the selected column from the input data.
E.g., "Electronic books".}

\item{denominator}{A character vector specifying a numeric statistic from the
ARL survey data to be used as the numerator. The vector should be
an exact match to the selected column from the input data.
E.g., "Total fulltime students".}

\item{members}{A character vector specifying up to five ARL member
institutes of interest, as identified in the dataset. E.g.,
c("BOSTON", "TORONTO", "OTTAWA", "LAVAL", "HARVARD").}

\item{years}{A numeric vector specifying up to 5 calendar years
for which data should be plotted, e.g., c(2015, 2016, 2017,
2018, 2019). If no value is provided (i.e., NA), then most
recent five years available in the uploaded data will be used.
If more than 5 values provided, last 5 values will be selected.
Default is NA.}
}
\value{
Returns bar plots showing varying ratios specified below:
\itemize{
  \item proFTETopPerFaculty - A barplot showing members with highest
        ratio of total library support staff counts (FTE) per
        teaching faculty, over user selected number of years.
  \item proFTETopPerStudent - A barplot showing members with highest
        ratio of total library support staff counts (FTE) per
        student (full-time, FT, and part-time, PT), over user selected
        number of years.
  \item proFTETopPerGradStudent - A barplot showing members with highest
        ratio of total library support staff counts (FTE) per graduate
        student (full-time, FT, and part-time, PT), over user selected
        number of years.
  \item proFTETopPerDoctoral - A barplot showing members with highest
        ratio of total library support staff counts (FTE) per doctoral
        degree awarded, over user selected number of years.
  \item proPerFacultyUserSelected - A barplot showing ratio of library
        support staff counts (FTE) per teaching faculty for user
        selected ARL members, over user selected number of years.
  \item proPerStudentUserSelected - A barplot showing ratio of library
        support staff counts (FTE) per student (full-time, FT, and
        part-time, PT) for user selected ARL members, over user selected
        number of years.
  \item proPerGradStudentUserSelected - A barplot showing ratio of total
        library support staff counts (FTE) per graduate student
        (full-time, FT, and part-time, PT) for user selected ARL members,
        over user selected number of years.
  \item proPerDoctoralUserSelected - A barplot showing ratio of total
        library expenditures per per doctoral degree awarded for user
        selected ARL members, over user selected number of years.
}
}
\description{
A function to build and visualize a custom ratio based on user
selected numerator and denominator from various statistics reported
in the annual survey of Association of Research Libraries (ARL). The
ratio can be visualized for user selected ARL members over user selected
years as bar plots.
}
\examples{
customRatioBuilder(dataARL = ARLDataDownload,
                   numerator = "Electronic books",
                   denominator = "Total fulltime students",
                   members = c("BOSTON", "TORONTO", "OTTAWA", "LAVAL", "HARVARD"),
                   years = c(2015, 2016, 2017, 2022, 2018, 2019))

}
