% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_shinyLibStats.R
\name{shinyLibStats}
\alias{shinyLibStats}
\title{Launch Shiny App For Package libraryStatistics}
\usage{
shinyLibStats()
}
\value{
Open up the shiny page.
}
\description{
A function that launches the shiny app for this package. The shiny app is a
dashboard that permit to visualize, track trends, and compare up to five years
of data downloaded directly from ARL data portal, with no data cleaning involved.
Upon uploading of data, the user is directed to select an institute of choice and
up to five years of choice. The user can then see the summary of dataset by
each year, and visually compare the selected institute against other libraries
for indicators defined by ARL. Additionally, useful ratios and percentages,
such as library expenditures per faculty, per student, and per doctoral
degree awarded, can be visualized and compared.
}
\examples{
\dontrun{
libraryStatistics::shinyLibStats()
}

}
\references{
Silva, A. et al. (2019). A multivariate Poisson-log normal mixture model
for clustering transcriptome sequencing data. \emph{BMC Bioinformatics} 20.
\href{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-019-2916-0}{Link}

Subedi, S., and R. Browne (2020). A parsimonious family of multivariate Poisson-lognormal
distributions for clustering multivariate count data. arXiv preprint arXiv:2004.06857.
\href{https://arxiv.org/pdf/2004.06857.pdf}{Link}
}
\author{
Anjali Silva, \email{a.silva@utoronto.ca}
}
