% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectionStats.R
\name{visCollectionData}
\alias{visCollectionData}
\title{A Bar Plot to Compare Collection Statistics}
\usage{
visCollectionData(dataARL, institute, years = NA)
}
\arguments{
\item{dataARL}{A data frame containing data downloaded from
ARL. The years should be placed along rows. The first column must
be 'Year', followed by other variables in no particular order,
e.g., 'Institution Name', 'Institution type', etc.}

\item{institute}{A character vector specifying the institute of
interest, as identified in the dataset. E.g., "TORONTO" for
University of Toronto Libraries.}

\item{years}{A numeric vector specifying up to 5 calendar years
for which data should be plotted, e.g., c(2015, 2016, 2017,
2018, 2019). If no value is provided (i.e., NA), then most
recent five years available in the data will be used. If more
than 5 values provided, last 5 values will be selected. Default
is NA.}
}
\value{
Returns three bar plots showing collection statistics
\itemize{
  \item InstCanadianPlot - A bar plot comparing Canadian institutes
        based on titles held.
  \item instTypePlot - A bar plot comparing maximum title holders by
        institute type. Types include: "Canadian", "Private", "State",
        and "Nonacademic".
  \item academicPlot - A bar plot comparing maximum title holders by
        academic institute type. Types include: "Canadian" and "State".
}
}
\description{
A function to visualize collection statistics. Institution types are
assumed to be of the categories: "Canadian", "Canadian Nonacademic",
"Private", "State", and "Nonacademic". For collection statistics
visualization, the following variables (or columns) are required
in the dataset: "Year", "Institution Name", "Institution type",
"Region", "Rank in ARL investment index", "ARL investment index value",
"Titles held", "Volumes held", and "Electronic books".
}
\examples{
visCollectionData(dataARL = ARLDataDownload,
                  institute = "TORONTO",
                  years = c(2015, 2016, 2017, 2018, 2019))

}
