% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBookStats.R
\name{viseBookData}
\alias{viseBookData}
\title{Plots to Compare eBooks in Collection Over Years}
\usage{
viseBookData(dataARL, institute, years = NA)
}
\arguments{
\item{dataARL}{A data frame containing data downloaded from
ARL. The years should be placed along rows. The first column must
be 'Year', followed by other variables in no particular order,
e.g., 'Institution Name', 'Institution type', etc.}

\item{institute}{A character vector specifying the institute of
interest, as identified in the dataset. E.g., "TORONTO" for
University of Toronto Libraries.}

\item{years}{A numeric vector specifying up to 5 calendar years
for which data should be plotted, e.g., c(2015, 2016, 2017,
2018, 2019). If no value is provided (i.e., NA), then most
recent five years available in the data will be used. If more
than 5 values provided, last 5 values will be selected. Default
is NA.}
}
\value{
Returns three bar plots showing volume statistics
\itemize{
  \item eBookUserInstitute - A lineplot comparing user selected
        institute over user selected number of years for eBooks
        held. The median line is provided for comparison.
  \item eBookVolumeComp - A stacked barplot showing total volumes
        held and eBook proportion over the period selected.
  \item eBookInstCanadian - A barplot comparing Canadian institutes
        based on eBooks, along with user selected institute over
        user selected number of years. The median is provided for
        comparison.
  \item eBookInstType - A barplot comparing maximum eBooks by
        institute type over user selected number of years. The user
        selected institute is provided for comparison. Institute
        types include: "Canadian", "Private", "State", and "Nonacademic".
        The median is provided for comparison.
  \item eBookAcademicPlot - A barplot comparing maximum eBooks by
        academic institute type over user selected number of years.
        The user selected institute is provided for comparison.
        Institute types include: "Canadian" and "State".
        The median is provided for comparison.
  \item eBookARLRankTop - A barplot comparing eBooks by
        top 5 ARL investment ranks over user selected number of years.
        The user selected institute is provided for comparison.
        The median is provided for comparison.
}
}
\description{
A function to visualize eBook statistics using multiple plot types
as described below. Institution types in the input data are
assumed to be of the categories: "Canadian", "Canadian Nonacademic",
"Private", "State", and "Nonacademic". For volume statistics
visualization, the following variables (or columns) are required
in the dataset: "Year", "Institution Name", "Institution type",
"Region", "Rank in ARL investment index", "ARL investment index value",
and "Electronic books".
}
\examples{
viseBookData(dataARL = ARLDataDownload,
             institute = "TEXAS STATE",
             years = c(2015, 2016, 2017, 2022, 2018, 2019))

}
