% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_profStaffCounts.R
\name{visProfStaffCounts}
\alias{visProfStaffCounts}
\title{Plots to Compare Professional Library Staff Counts Over Years}
\usage{
visProfStaffCounts(dataARL, members, years = NA)
}
\arguments{
\item{dataARL}{A dataframe containing ARL survey data directly
downloaded from ARL platform. The years should be placed along
rows. The first column must be 'Year', followed by other variables
in no particular order, e.g., 'Institution Name', 'Institution type',
etc. To download data from ARL Data Portal, it is recommended that
all variables are selected, with columns being 'Variables' and data
sorted by 'Institution Name' (default options).}

\item{members}{A character vector specifying up to five ARL member
institutes of interest, as identified in the dataset. E.g.,
c("BOSTON", "TORONTO", "OTTAWA", "LAVAL", "HARVARD").}

\item{years}{A numeric vector specifying up to 5 calendar years
for which data should be plotted, e.g., c(2015, 2016, 2017,
2018, 2019). If no value is provided (i.e., NA), then most
recent five years available in the uploaded data will be used.
If more than 5 values provided, last 5 values will be selected.
Default is NA.}
}
\value{
Returns bar plots showing varying ratios specified below:
\itemize{
  \item proFTETopPerFaculty - A barplot showing members with highest
        ratio of total library professional staff counts (FTE) per
        teaching faculty, over user selected number of years.
  \item proFTETopPerStudent - A barplot showing members with highest
        ratio of total library professional staff counts (FTE) per
        student (full-time, FT, and part-time, PT), over user selected
        number of years.
  \item proFTETopPerGradStudent - A barplot showing members with highest
        ratio of total library professional staff counts (FTE) per graduate
        student (full-time, FT, and part-time, PT), over user selected
        number of years.
  \item proFTETopPerDoctoral - A barplot showing members with highest
        ratio of total library professional staff counts (FTE) per doctoral
        degree awarded, over user selected number of years.
  \item proPerFacultyUserSelected - A barplot showing ratio of library
        professional staff counts (FTE) per teaching faculty for user
        selected ARL members, over user selected number of years.
  \item proPerStudentUserSelected - A barplot showing ratio of library
        professional staff counts (FTE) per student (full-time, FT, and
        part-time, PT) for user selected ARL members, over user selected
        number of years.
  \item proPerGradStudentUserSelected - A barplot showing ratio of total
        library professional staff counts (FTE) per graduate student
        (full-time, FT, and part-time, PT) for user selected ARL members,
        over user selected number of years.
  \item proPerDoctoralUserSelected - A barplot showing ratio of total
        library expenditures per per doctoral degree awarded for user
        selected ARL members, over user selected number of years.
}
}
\description{
A function to visualize library professional staff counts, full-time
equivalent (FTE), as ratios in comparison to various statistics reported
in the annual survey of Association of Research Libraries (ARL) as bar
plots. This is question 13a on ARL survey.
}
\examples{
visProfStaffCounts(dataARL = ARLDataDownload,
                   members = c("BOSTON", "TORONTO", "OTTAWA", "LAVAL", "HARVARD"),
                   years = c(2015, 2016, 2017, 2022, 2018, 2019))

}
