% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffCounts.R
\name{visStaffCounts}
\alias{visStaffCounts}
\title{Plots to Compare Library Staff Counts Over Years}
\usage{
visStaffCounts(dataARL, members, years = NA)
}
\arguments{
\item{dataARL}{A dataframe containing ARL survey data directly
downloaded from ARL platform. The years should be placed along
rows. The first column must be 'Year', followed by other variables
in no particular order, e.g., 'Institution Name', 'Institution type',
etc. To download data from ARL Data Portal, it is recommended that
all variables are selected, with columns being 'Variables' and data
sorted by 'Institution Name' (default options).}

\item{members}{A character vector specifying up to five ARL member
institutes of interest, as identified in the dataset. E.g.,
c("BOSTON", "TORONTO", "OTTAWA", "LAVAL", "HARVARD").}

\item{years}{A numeric vector specifying up to 5 calendar years
for which data should be plotted, e.g., c(2015, 2016, 2017,
2018, 2019). If no value is provided (i.e., NA), then most
recent five years available in the uploaded data will be used.
If more than 5 values provided, last 5 values will be selected.
Default is NA.}
}
\value{
Returns three bar plots showing staff count statistics
\itemize{
  \item staffFTEUserInstitute - A lineplot comparing user selected
        institute over user selected number of years for staff counts.
        The median line is provided for comparison.
}
}
\description{
A function to visualize library professional staff counts, full-time
equivalent (FTE), as ratios in comparison to various statistics reported
in the annual survey of Association of Research Libraries (ARL) as bar plots.
}
\examples{
visStaffCounts(dataARL = ARLDataDownload,
               members = c("BOSTON", "TORONTO", "OTTAWA", "LAVAL", "HARVARD"),
               years = c(2015, 2016, 2017, 2022, 2018, 2019))

}
