% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStats.R
\name{visSummaryAllData}
\alias{visSummaryAllData}
\title{Plots to Compare Summary of Regions}
\usage{
visSummaryAllData(dataARL, institute, years = NA)
}
\arguments{
\item{dataARL}{A data frame containing data downloaded from
ARL. The years should be placed along rows. The first column must
be 'Year', followed by other variables in no particular order,
e.g., 'Institution Name', 'Institution type', etc.}

\item{institute}{A character vector specifying the institute of
interest, as identified in the dataset. E.g., "TORONTO" for
University of Toronto Libraries.}

\item{years}{A numeric vector specifying up to 5 calendar years
for which data should be plotted, e.g., c(2015, 2016, 2017,
2018, 2019). If no value is provided (i.e., NA), then most
recent five years available in the data will be used. If more
than 5 values provided, last 5 values will be selected. Default
is NA.}
}
\value{
Returns three bar plots showing title statistics
\itemize{
  \item titleUserInstitute - A lineplot comparing user selected
        institute over user selected number of years for titles
        held. The median line is provided for comparison.
  \item titleInstCanadian - A barplot comparing Canadian institutes
        based on titles held, along with user selected institute over
        user selected number of years. The median is provided for
        comparison.
  \item titleInstType - A barplot comparing maximum title by
        institute type over user selected number of years. The user
        selected institute is provided for comparison. Institute
        types include: "Canadian", "Private", "State", and "Nonacademic".
         The median is provided for comparison.
  \item titleAcademicPlot - A barplot comparing maximum title by
        academic institute type over user selected number of years.
        The user selected institute is provided for comparison.
        Institute types include: "Canadian" and "State". The median
        is provided for comparison.
  \item titleARLRankTop - A barplot comparing titles by
        top 5 ARL investment ranks over user selected number of years.
        The user selected institute is provided for comparison.
        The median is provided for comparison.
}
}
\description{
A function to visualize sample sizes on institutes submitting data to ARL for
each year via barplots for the entire dataset uploaded by the user. Institution
types in the input data are assumed to be of the categories: "Canadian",
"Canadian Nonacademic", "Private", "State", and "Nonacademic". For summary
visualization, the following variables (or columns) are required
in the dataset: "Year", "Institution type", "Region", and "Year".
}
\examples{
visSummaryAllData(dataARL = ARLDataDownload,
                years = c(2015, 2016, 2017, 2022, 2018, 2019))

}
