% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totLibExpStats.R
\name{visTotalLibraryExp}
\alias{visTotalLibraryExp}
\title{Plots to Compare Total Library Expenditures Over Years}
\usage{
visTotalLibraryExp(dataARL, members = NA, years = NA)
}
\arguments{
\item{dataARL}{A dataframe containing ARL survey data directly
downloaded from ARL platform. The years should be placed along
rows. The first column must be 'Year', followed by other variables
in no particular order, e.g., 'Institution Name', 'Institution type',
etc.}

\item{members}{A character vector specifying up to five ARL member
institutes of interest, as identified in the dataset. E.g.,
c("BOSTON", "TORONTO", "OTTAWA", "LAVAL", "HARVARD").}

\item{years}{A numeric vector specifying up to 5 calendar years
for which data should be plotted, e.g., c(2015, 2016, 2017,
2018, 2019). If no value is provided (i.e., NA), then most
recent five years available in the uploaded data will be used.
If more than 5 values provided, last 5 values will be selected.
Default is NA.}
}
\value{
Returns three bar plots showing title statistics
\itemize{
  \item tleTopPerFaculty - A barplot comparing user selected institute
        with institutes with highest value for total library
        expenditures per teaching faculty, over user selected number of years.
  \item tleTopPerStudent - A barplot comparing user selected institute
        with institutes with highest value for total library
        expenditures per total students reported that are either part-time (PT)
        or full-time (FT), over user selected number of years.
  \item tleTopPerGradStudent - A barplot comparing user selected institute
        with institutes with highest value for total library
        expenditures per total graduate students reported that are either
        part-time (PT) or full-time (FT), over user selected number of years.
  \item tleTopPerDoctoral - A barplot comparing user selected institute
        with institutes with highest value for total library
        expenditures per doctoral degrees awarded, over user selected number
        of years.
}
}
\description{
A function to visualize total library expenditures in United
States Dollars (USD) as ratios in comparison to various statistics
reported in the annual survey of Association of Research Libraries (ARL).
}
\examples{
visTotalLibraryExp(dataARL = ARLDataDownload,
                   members = c("BOSTON", "TORONTO", "OTTAWA", "LAVAL", "HARVARD"),
                   years = c(2015, 2016, 2017, 2022, 2018, 2019))

}
