% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staffCounts.R
\name{visStaffCounts}
\alias{visStaffCounts}
\title{Plots to Compare Library Staff Counts Over Years}
\usage{
visStaffCounts(dataARL, institute, years = NA)
}
\arguments{
\item{dataARL}{A data frame containing data downloaded from
ARL. The years should be placed along rows. The first column must
be 'Year', followed by other variables in no particular order,
e.g., 'Institution Name', 'Institution type', etc.}

\item{institute}{A character vector specifying the institute of
interest, as identified in the dataset. E.g., "TORONTO" for
University of Toronto Libraries.}

\item{years}{A numeric vector specifying up to 5 calendar years
for which data should be plotted, e.g., c(2015, 2016, 2017,
2018, 2019). If no value is provided (i.e., NA), then most
recent five years available in the data will be used. If more
than 5 values provided, last 5 values will be selected. Default
is NA.}
}
\value{
Returns three bar plots showing staff count statistics
\itemize{
  \item staffFTEUserInstitute - A lineplot comparing user selected
        institute over user selected number of years for staff counts.
        The median line is provided for comparison.
  \item staffFTEInstCanadian - A barplot comparing Canadian institutes
        based on staff counts, along with user selected institute over
        user selected number of years. The median is provided for
        comparison.
  \item staffFTEInstType - A barplot comparing maximum staff counts by
        institute type over user selected number of years. The user
        selected institute is provided for comparison. Institute
        types include: "Canadian", "Private", "State", and "Nonacademic".
        The median is provided for comparison.
  \item staffFTEAcademicPlot - A barplot comparing maximum staff counts by
        academic institute type over user selected number of years.
        The user selected institute is provided for comparison.
        Institute types include: "Canadian" and "State".
        The median is provided for comparison.
  \item staffFTEARLRankTop - A barplot comparing staff counts by
        top 5 ARL investment ranks over user selected number of years.
        The user selected institute is provided for comparison.
        The median is provided for comparison.
   \item staffFTEComp - A barplot showing proportion of professional,
        support and student assitant staff, making up the total
        staff (FTE) counts in the selected institute over the
        user provided period.
   \item staffFTEperFaculty - A barplot showing the proportion of
        professional staff (FTE) per full-time instructional faculty.
   \item staffFTEperStudent - A barplot showing the proportion of
        professional staff (FTE) per students of status full-time
        and part-time, both graduate and undergraduate.
   \item staffFTEperGradStudent - A barplot showing the proportion of
        professional staff (FTE) per graduate students of status full-time
        and part-time.
   \item staffFTEperDoctoral - A barplot showing the proportion of
        professional staff (FTE) per doctoral degree awarded.
   \item profStaffPercentage - A barplot showing the proportion of
        professional staff (FTE) as a percentage of total staff.
   \item profStaffAllData - A violin plot showing the distribution
        of professional staff counts per years selected by user, for the
        entire dataset uploaded.
   \item staffAllData - A violin plot showing the distribution
        of all (professional, support and casual) staff counts per
        years selected by user, for the entire dataset uploaded.
}
}
\description{
A function to visualize library staff counts, full time equivalent (FTE)
using multiple plot types as described below. Institution types in the
input data are assumed to be of the categories: "Canadian",
"Canadian Nonacademic", "Private", "State", and "Nonacademic". For
staff count visualization, the following variables (or columns) are
required in the dataset: "Year", "Institution Name", "Institution type",
"Region", "Rank in ARL investment index", "ARL investment index value",
"Professional staff", "Support staff", "Student assistants", and
"Total prof. + support + student staff".
}
\examples{
visStaffCounts(dataARL = ARLDataDownload,
               institute = "TEXAS STATE",
               years = c(2015, 2016, 2017, 2022, 2018, 2019))

}
