% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totLibExpStats.R
\name{visTotalLibraryExp}
\alias{visTotalLibraryExp}
\title{Plots to Compare Total Library Expenditures Over Years}
\usage{
visTotalLibraryExp(dataARL, institute, years = NA)
}
\arguments{
\item{dataARL}{A data frame containing data downloaded from
ARL. The years should be placed along rows. The first column must
be 'Year', followed by other variables in no particular order,
e.g., 'Institution Name', 'Institution type', etc.}

\item{institute}{A character vector specifying the institute of
interest, as identified in the dataset. E.g., "TORONTO" for
University of Toronto Libraries.}

\item{years}{A numeric vector specifying up to 5 calendar years
for which data should be plotted, e.g., c(2015, 2016, 2017,
2018, 2019). If no value is provided (i.e., NA), then most
recent five years available in the data will be used. If more
than 5 values provided, last 5 values will be selected. Default
is NA.}
}
\value{
Returns three bar plots showing title statistics
\itemize{
  \item tleAllData - Violin plots showing the distribution
        of total library expenditures for all institutes in dataset
        uploaded by user for the years chosen by the user.
  \item tleUserInstitute - A lineplot comparing user selected
        institute over user selected number of years for total
        library expenditures. The median line is provided for
        comparison.
  \item tleExpComp - A barplot showing the proportion making up total
        library expenditures for user selected institute over user selected
        number of years. The three categories include total materials
        expenditures, total salaries and wages, and other operating expenditures.
  \item tleInstCanadian - A barplot comparing Canadian institutes
        based on total library expenditure, along with user selected
        institute over user selected number of years. The median is
        provided for comparison.
  \item tleInstType - A barplot comparing maximum total library expenditure
        by institute type over user selected number of years. The user
        selected institute is provided for comparison. Institute
        types include: "Canadian", "Private", "State", and "Nonacademic".
         The median is provided for comparison.
  \item tleAcademicPlot - A barplot comparing maximum total library
        expenditure by academic institute type over user selected number
        of years. The user selected institute is provided for comparison.
        Institute types include: "Canadian" and "State". The median
        is provided for comparison.
  \item tleARLRankTop - A barplot comparing total library expenditures by
        top 5 ARL investment ranks over user selected number of years.
        The user selected institute is provided for comparison.
        The median is provided for comparison.
  \item tleARLRankTopPerFaculty - A barplot comparing user selected institute
        with institutes with high ARL investment ranks, for total library
        expenditures per teaching faculty, over user selected number of years.
  \item tleARLRankTopPerStudent - A barplot comparing user selected institute
        with institutes with high ARL investment ranks, for total library
        expenditures per total students reported that are either part-time (PT)
        or full-time (FT), over user selected number of years.
  \item tleARLRankTopPerGradStudent - A barplot comparing user selected institute
        with institutes with high ARL investment ranks, for total library
        expenditures per total graduate students reported that are either
        part-time (PT) or full-time (FT), over user selected number of years.
  \item tleARLRankTopPerDoctoral - A barplot comparing user selected institute
        with institutes with high ARL investment ranks, for total library
        expenditures per doctoral degrees awarded, over user selected number
        of years.
  \item tleTop - A barplot comparing user selected institute with institutes
        with highest value for total library expenditures, over user selected
        number of years.
  \item tleTopPerFaculty - A barplot comparing user selected institute
        with institutes with highest value for total library
        expenditures per teaching faculty, over user selected number of years.
  \item tleTopPerStudent - A barplot comparing user selected institute
        with institutes with highest value for total library
        expenditures per total students reported that are either part-time (PT)
        or full-time (FT), over user selected number of years.
  \item tleTopPerGradStudent - A barplot comparing user selected institute
        with institutes with highest value for total library
        expenditures per total graduate students reported that are either
        part-time (PT) or full-time (FT), over user selected number of years.
  \item tleTopPerDoctoral - A barplot comparing user selected institute
        with institutes with highest value for total library
        expenditures per doctoral degrees awarded, over user selected number
        of years.
}
}
\description{
A function to visualize total library expenditures in United
States Dollars (USD) using multiple plot types as described below.
Institution types in the input data are assumed to be of the
categories: "Canadian", "Canadian Nonacademic",
"Private", "State", and "Nonacademic". For title statistics
visualization, the following variables (or columns) are required
in the dataset: "Year", "Institution Name", "Institution type",
"Region", "Rank in ARL investment index", "ARL investment index value",
and "Total library expenditures".
}
\examples{
visTotalLibraryExp(dataARL = ARLDataDownload,
                   institute = "BOSTON",
                   years = c(2015, 2016, 2017, 2022, 2018, 2019))

}
