% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_totLibExpStats.R
\name{visTotalLibraryExp}
\alias{visTotalLibraryExp}
\title{Plots to Compare Total Library Expenditures Over Years}
\usage{
visTotalLibraryExp(dataARL, members = NA, years = NA)
}
\arguments{
\item{dataARL}{A dataframe containing ARL survey data directly
downloaded from ARL platform. The years should be placed along
rows. The first column must be 'Year', followed by other variables
in no particular order, e.g., 'Institution Name', 'Institution type',
etc. To download data from ARL Data Portal, it is recommended that
all variables are selected, with columns being 'Variables' and data
sorted by 'Institution Name' (default options).}

\item{members}{A character vector specifying up to five ARL members
of interest, as identified in the dataset. E.g.,
c("BOSTON", "TORONTO", "OTTAWA", "LAVAL", "HARVARD").}

\item{years}{A numeric vector specifying up to 5 calendar years
for which data should be plotted, e.g., c(2015, 2016, 2017,
2018, 2019). If no value is provided (i.e., NA), then most
recent five years available in the uploaded data will be used.
If more than 5 values provided, last 5 values will be selected.
Default is NA.}
}
\value{
Returns bar plots showing varying ratios specified below:
\itemize{
  \item tleTopPerFaculty - A barplot showing ARL members with highest ratio of total
        library expenditures per teaching faculty, over user selected number of years.
  \item tleTopPerStudent - A barplot showing ARL members with highest ratio of total
        library expenditures per student (full-time, FT, and part-time, PT), over
         user selected number of years.
  \item tleTopPerGradStudent - A barplot showing ARL members with highest ratio of total
        library expenditures per graduate student (full-time, FT, and part-time, PT),
        over user selected number of years.
  \item tleTopPerDoctoral - A barplot showing ARL members with highest ratio of total
        library expenditures per doctoral degree awarded, over user selected number
        of years.
  \item tlePerFacultyUserSelected - A barplot showing ratio of total library expenditures per
        teaching faculty for user selected ARL members, over user selected number of years.
  \item tlePerStudentUserSelected - A barplot showing ratio of total library expenditures per
        student (full-time, FT, and part-time, PT) for user selected ARL members, over user
        selected number of years.
  \item tlePerGradStudentUserSelected - A barplot showing ratio of total library expenditures per
        graduate student (full-time, FT, and part-time, PT) for user selected ARL members, over user
        selected number of years.
  \item tleTopPerDoctoralUserSelected - A barplot showing ratio of total library expenditures per
        per doctoral degree awarded for user selected ARL members, over user
        selected number of years.
}
}
\description{
A function to visualize total library expenditures in United
States Dollars (USD) as ratios in comparison to various statistics
reported in the annual survey of Association of Research Libraries (ARL)
as bar plots. Note, this function provides question 6 of ARL survey
as the numerator of the ratio.
}
\examples{
visTotalLibraryExp(dataARL = ARLDataDownload,
                   members = c("BOSTON", "TORONTO", "OTTAWA", "LAVAL", "HARVARD"),
                   years = c(2015, 2016, 2017, 2022, 2018, 2019))

}
\references{
Mian, A., & Gross, H. (2023). ARL Statistics 2022. Washington, DC:
Association of Research Libraries.
\href{https://publications.arl.org/ARL-Statistics-2022/}{Link}

Association of Research Libraries. (2023). ARL Statistics 2023
Instructions. \href{https://www.arlstatistics.org/resources/stats_instructions}{Link}
}
\author{
{Anjali Silva, \email{a.silva@utoronto.ca}, Klara Maidenberg, \email{klara.maidenberg@utoronto.ca}}
}
